Add USER_NEW trigger handler, split handlers code
